requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
-e git+https://github.com/jendrikseipp/biblib.git@v0.1.1#egg=biblib
    # via -r requirements.in
anyio==4.4.0
    # via watchfiles
blinker==1.8.2
    # via pelican
docutils==0.21.2
    # via pelican
exceptiongroup==1.2.1
    # via anyio
feedgenerator==2.1.0
    # via pelican
idna==3.7
    # via anyio
invoke==2.2.0
    # via -r requirements.in
jinja2==3.1.4
    # via pelican
livereload==2.7.0
    # via -r requirements.in
markdown==3.3.6
    # via pelican
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.1
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
ordered-set==4.1.0
    # via pelican
pelican[markdown]==4.9.1
    # via -r requirements.in
pygments==2.18.0
    # via
    #   pelican
    #   rich
python-dateutil==2.8.2
    # via pelican
pytz==2022.1
    # via feedgenerator
pyyaml==6.0.1
    # via -r requirements.in
rich==13.7.1
    # via pelican
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
tornado==6.1
    # via livereload
typing-extensions==4.2.0
    # via anyio
unidecode==1.3.8
    # via pelican
watchfiles==0.22.0
    # via pelican
